+++ lib/api/registry/person/get.ex
@@ -1,17 +1,17 @@
-defmodule Api.Location.Get do
+defmodule Api.Registry.Person.Get do
   @moduledoc false
-  alias Api.Location.Country
+  alias Api.Registry.Person
   alias Api.Repo
 
   @doc false
   def call(id) do
-    Country
+    Person
     |> Repo.get!(id)
     |> then(&{:ok, &1})
   rescue
     _ ->
       :id
-      |> Country.invalid_changeset(id, "not found")
+      |> Person.invalid_changeset(id, "not found")
       |> then(&{:error, &1})
   end
 end
