+++ lib/api/user_management/user/delete.ex
@@ -1,4 +1,4 @@
-defmodule Api.UserManagement.Delete do
+defmodule Api.UserManagement.User.Delete do
   @moduledoc false
   alias Api.Repo
   alias Api.UserManagement.User
